chore: split command creation from run
